[Rename] common header file
